Extract chart.js stuff into a dedicated component
