+++ Spbs.Ui/Features/Visualization/Graphs/GraphComponent.razor
@@ -1,0 +1,4 @@
+@typeparam TDataItem
+
+
+<canvas id="@_canvasId" />
